feat: refactor navigation components to use centralized formatting functions
