+++ mobile/app/(tabs)/maps/index.tsx
@@ -67,8 +67,6 @@ export default function MapsScreen() {
     error: routeError,
     calculateRoute,
     clearRoute,
-    formatDistance,
-    formatDuration,
   } = useRouteCalculation({ mapRef });
 
   // Ref to store stopNavigation to avoid circular dependency
@@ -95,8 +93,6 @@ export default function MapsScreen() {
     startNavigation,
     stopNavigation,
     handleLocationUpdate,
-    formatDistanceMeters,
-    formatTimeSeconds,
     turnInstructions,
   } = useNavigationManagement({
     mapRef,
@@ -311,15 +307,9 @@ export default function MapsScreen() {
               progress={navigationState.progress}
               turnInstructions={turnInstructions}
               onStopNavigation={handleStopNavigation}
-              formatDistanceMeters={formatDistanceMeters}
-              formatTimeSeconds={formatTimeSeconds}
             />
           ) : (
-            <RouteInfo
-              route={route}
-              formatDistance={formatDistance}
-              formatDuration={formatDuration}
-            />
+            <RouteInfo route={route} />
           )}
         </BottomSheetView>
       </BottomSheet>
