+++ mobile/components/maps/route-info.tsx
@@ -4,15 +4,14 @@ import { Navigation, Clock } from 'lucide-react-native';
 import { useTheme } from '@react-navigation/native';
 import { Text } from '@/components/ui/text';
 import { TurnByTurnList } from '@/components/maps/turn-by-turn-list';
+import { formatDistanceMeters, formatTimeSeconds } from '@/utils/formatting';
 import { useMemo } from 'react';
 
 interface RouteInfoProps {
   route: Route | null;
-  formatDistance: (route: Route) => string;
-  formatDuration: (route: Route) => string;
 }
 
-export function RouteInfo({ route, formatDistance, formatDuration }: RouteInfoProps) {
+export function RouteInfo({ route }: RouteInfoProps) {
   const { colors } = useTheme();
 
   const turnInstructions = useMemo(() => {
@@ -22,24 +21,6 @@ export function RouteInfo({ route, formatDistance, formatDuration }: RouteInfoPr
     }
     return [];
   }, [route]);
-
-  const formatDistanceMeters = (meters: number): string => {
-    if (isNaN(meters) || meters < 0) return '0 m';
-    if (meters < 1000) {
-      return `${Math.round(meters)} m`;
-    }
-    return `${(meters / 1000).toFixed(1)} km`;
-  };
-
-  const formatTimeSeconds = (seconds: number): string => {
-    if (isNaN(seconds) || seconds < 0) return '0 min';
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-    if (hours > 0) {
-      return `${hours}h ${minutes}min`;
-    }
-    return `${minutes} min`;
-  };
 
   if (!route) return null;
 
@@ -51,7 +32,9 @@ export function RouteInfo({ route, formatDistance, formatDuration }: RouteInfoPr
           <Navigation color={colors.primary} size={20} />
           <View>
             <Text className="text-xs text-muted-foreground">Distance</Text>
-            <Text className="text-base font-semibold text-foreground">{formatDistance(route)}</Text>
+            <Text className="text-base font-semibold text-foreground">
+              {formatDistanceMeters(route.distance)}
+            </Text>
           </View>
         </View>
 
@@ -60,7 +43,9 @@ export function RouteInfo({ route, formatDistance, formatDuration }: RouteInfoPr
           <Clock color={colors.primary} size={20} />
           <View>
             <Text className="text-xs text-muted-foreground">Duration</Text>
-            <Text className="text-base font-semibold text-foreground">{formatDuration(route)}</Text>
+            <Text className="text-base font-semibold text-foreground">
+              {formatTimeSeconds(route.duration)}
+            </Text>
           </View>
         </View>
       </View>
@@ -69,11 +54,7 @@ export function RouteInfo({ route, formatDistance, formatDuration }: RouteInfoPr
       {turnInstructions.length > 0 && (
         <View>
           <Text className="mb-2 text-sm font-semibold">Steps</Text>
-          <TurnByTurnList
-            turnInstructions={turnInstructions}
-            formatDistanceMeters={formatDistanceMeters}
-            formatTimeSeconds={formatTimeSeconds}
-          />
+          <TurnByTurnList turnInstructions={turnInstructions} />
         </View>
       )}
     </View>
